ref: add service method and tests
